Add a temporary API DNS entry
The OpenShift deployment currently fails, because openshift-ansible
makes requests againtst the API (i.e. the load balancer).

This will now always fail without a manual intervention, because the
external load balancer won't have the master entries.

So this adds a DNS entry mapping the load balancer hostname to the
floating IP address of the first master node.

That way, the installation will succeed, and we can update the load
balancer after the deployment has finished.